feat(#31): middleware checkAdmin hecho

+++ routes/api.php
@@ -17,75 +17,61 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
+// ADMINISTRADOR
+//RESUMENES
+Route::get('/auth/numeroUsuarios', 'AuthController@getNumeroUsuarios')->middleware('checkAdmin');
+Route::get('/carrito/numeroCompras', 'CarritoController@NumeroCompras')->middleware('checkAdmin');
+Route::get('/carrito/resumenIngresos', 'CarritoController@ResumenIngresos')->middleware('checkAdmin');
+Route::get('/productos/productosMasComprados', 'ProductosController@productosMasComprados')->middleware('checkAdmin');
+Route::get('/productos/numeroProductos', 'ProductosController@NumeroProductos')->middleware('checkAdmin');
+Route::get('/recetas/numeroRecetas', 'RecetasController@NumeroRecetas')->middleware('checkAdmin');
+//CRUD
+Route::post('/redessociales', 'RedesSocialesController@CrearRedSocial')->middleware('checkAdmin');
+Route::delete('/redessociales/{id}', 'RedesSocialesController@BorrarRedSocial')->middleware('checkAdmin');
+Route::put('/redessociales/{id}', 'RedesSocialesController@ModificarRedSocial')->middleware('checkAdmin');
+Route::post('/productos/nuevo', 'ProductosController@anyadirProducto')->middleware('checkAdmin');
+Route::delete('/productos/{idProducto}', 'ProductosController@BorrarProducto')->middleware('checkAdmin');
+Route::patch('/productos/{idProducto}', 'ProductosController@ModificarProducto')->middleware('checkAdmin');
+Route::post('/categorias/nueva', 'CategoriasController@anyadirCategoria')->middleware('checkAdmin');
+Route::delete('/categorias/{categoria}', 'CategoriasController@BorrarCategoria')->middleware('checkAdmin');
+Route::post('/recetas/nueva', 'RecetasController@anyadirReceta')->middleware('checkAdmin');
+Route::delete('/recetas/{idReceta}', 'RecetasController@BorrarReceta')->middleware('checkAdmin');
+Route::get('/categorias/productosCategoria', 'CategoriasController@ProductosCategoria')->middleware('checkAdmin');
+
 //CATEGORIAS
 Route::get('/categorias', 'CategoriasController@GetAll');
-
-Route::get('/categorias/productosCategoria', 'CategoriasController@ProductosCategoria');
-
-Route::post('/categorias/nueva', 'CategoriasController@anyadirCategoria');
-
 Route::get('/categorias/{categoria}', 'CategoriasController@Get');
-
 Route::get('/categorias/{categoria}/productos', 'CategoriasController@ProductosPorCategoria');
-
-Route::delete('/categorias/{categoria}', 'CategoriasController@BorrarCategoria');
 
 //PRODUCTOS
 Route::get('/productos', 'ProductosController@GetAll');
-
-Route::get('/productos/numeroProductos', 'ProductosController@NumeroProductos');
-
-Route::get('/productos/productosMasComprados', 'ProductosController@productosMasComprados');
-
-Route::post('/productos/nuevo', 'ProductosController@anyadirProducto');
-
 Route::get('/productos/{producto}', 'ProductosController@Get');
-
 Route::get('/busqueda={producto}', 'ProductosController@GetBarra');
-
-Route::delete('/productos/{idProducto}', 'ProductosController@BorrarProducto');
 
 //RECETAS
 Route::get('/recetas', 'RecetasController@GetAll');
 
-Route::get('/recetas/numeroRecetas', 'RecetasController@NumeroRecetas');
-
-Route::post('/recetas/nueva', 'RecetasController@anyadirReceta');
-
-Route::delete('/recetas/{idReceta}', 'RecetasController@BorrarReceta');
-
 //CAROUSEL
 // Route::get('/carousel', 'CarouselController@GetAll');
-
 Route::get('/carousel', 'CarouselController@GetImagenesPromocion');
-
-//USUARIOS
 
 //REDES SOCIALES
 Route::get('/redessociales', 'RedesSocialesController@GetAll');
-
-Route::post('/redessociales', 'RedesSocialesController@CrearRedSocial');
-
 Route::get('/redessociales/{id}', 'RedesSocialesController@Get');
-Route::delete('/redessociales/{id}', 'RedesSocialesController@BorrarRedSocial');
-Route::put('/redessociales/{id}', 'RedesSocialesController@ModificarRedSocial');
 
 // CARRITO
-Route::get('/carrito/numeroCompras', 'CarritoController@NumeroCompras');
-Route::get('/carrito/resumenIngresos', 'CarritoController@ResumenIngresos');
 Route::get('/carrito/{idUsuario}', 'CarritoController@Get');
 Route::get('/historialCarritos/{idUsuario}', 'CarritoController@HistorialCarritos');
 Route::post('/insertarCarritoTemporal', 'CarritoController@InsertarCarrito');
 Route::post('/comprarCarrito', 'CarritoController@ComprarCarrito');
 
 //LOGIN
-
 Route::group([
     'middleware' => 'api',
     'prefix' => 'auth',
 ], function () {
+    // Route::get('numeroUsuarios', 'AuthController@getNumeroUsuarios');
     Route::post('me', 'AuthController@me');
-    Route::get('numeroUsuarios', 'AuthController@getNumeroUsuarios');
     Route::post('login', 'AuthController@login');
     Route::post('logout', 'AuthController@logout');
     Route::post('refresh', 'AuthController@refresh');
